Fix navigation
the `pages/NAVIGATION.md` referenced `cv/index.md` directly. Apparently,
one needs to refer to the directory `cv/` so that the `NAVIGATION.md` is
being read.

Also fixed a warning about the nav not referencing one of the .md files
in `pages` which is now inlined. The content of that file is probably
wrong/outdated but at least everything builds again.