Drop headers
the "author" field is somewhat misleading since most pages are not
entirely static lots of authors have in fact written a page. Since we
have an About page, we do not need individual authors, I think.

The title section in the header is probably too annoying to maintain.
We should just use the first header "#" of the markdown file instead
(the default.)