Revert "Add link() macro to create relative links from absolute paths"
This reverts commit 1de5c687cf6ebda5fafe67cdef2935af6340d3bc.

We probably can just assume that we will always deploy to / and do not
have to worry about relative links here at all. Anyway, I wanted this
(overengineered piece of code) to be in the history in case we want to
support deployment to different base paths at some point.